+++ index.js
@@ -1,4 +1,5 @@
 import { Validator } from './src/constant.js';
+import { addValidator } from './src/getValidator.js';
 import { validate } from './src/validate.js';
 
-export { validate, Validator };
+export { validate, addValidator, Validator };