+++ doc/CHANGELOG.md
@@ -1,3 +1,8 @@
+v. 0.2.0
+
+- RegEx Validator
+- Custom Validator support
+
 v. 0.1.1
 
 - fixed import extensions